// src/app/models/pricing.interface.ts
export interface PricingI {
    id?: string;
    title: string;
    price: string;
    characteristic1: string;
    characteristic2: string;
    characteristic3: string;
    characteristic4?: string;
    characteristic5?: string;
    characteristic6?: string;
    characteristic7?: string;
    characteristic8?: string;
}